Moving transitionJiraIssue to a better home.
